backend/requirements.txt
python-dotenv==1.0.1
langchain==0.3.13
google-genai==0.3.0
fastapi==0.112.0
python-docx==1.1.2
Spire.Doc==12.12.0
crewai==0.98.0
crewai-tools==0.17.0
uvicorn==0.30.5
